client: Move attestation code to its own file
